исправления по экрану категории
